moved component props interfaces to new folder
