+++ source/battle-board/battle-board/models/component-props/button-data.ts
