+++ source/battle-board/battle-board/models/component-props/search-bar-data.ts
